Add details to XML describing why a Run was skipped
