+++ internal/command/junit/junit.go
@@ -81,7 +81,7 @@ func (v *TestJUnitXMLFile) Save(suite *moduletest.Suite) tfdiags.Diagnostics {
 
 	// Prepare XML content
 	sources := v.configLoader.Parser().Sources()
-	xmlSrc, err := junitXMLTestReport(suite, sources)
+	xmlSrc, err := junitXMLTestReport(suite, v.testSuiteRunner.IsStopped(), sources)
 	if err != nil {
 		diags = diags.Append(&hcl.Diagnostic{
 			Severity: hcl.DiagError,
@@ -142,7 +142,7 @@ type testCase struct {
 	Timestamp string  `xml:"timestamp,attr,omitempty"`
 }
 
-func junitXMLTestReport(suite *moduletest.Suite, sources map[string][]byte) ([]byte, error) {
+func junitXMLTestReport(suite *moduletest.Suite, suiteRunnerStopped bool, sources map[string][]byte) ([]byte, error) {
 	var buf bytes.Buffer
 	enc := xml.NewEncoder(&buf)
 	enc.EncodeToken(xml.ProcInst{
@@ -194,7 +194,7 @@ func junitXMLTestReport(suite *moduletest.Suite, sources map[string][]byte) ([]b
 			},
 		})
 
-		for _, run := range file.Runs {
+		for i, run := range file.Runs {
 			// Each run is a "test case".
 
 			testCase := testCase{
@@ -213,9 +213,10 @@ func junitXMLTestReport(suite *moduletest.Suite, sources map[string][]byte) ([]b
 			}
 			switch run.Status {
 			case moduletest.Skip:
+				message, body := getSkipDetails(i, file, suiteRunnerStopped)
 				testCase.Skipped = &withMessage{
-					// FIXME: Is there something useful we could say here about
-					// why the test was skipped?
+					Message: message,
+					Body:    body,
 				}
 			case moduletest.Fail:
 				testCase.Failure = &withMessage{
@@ -258,6 +259,37 @@ func junitXMLTestReport(suite *moduletest.Suite, sources map[string][]byte) ([]b
 	enc.EncodeToken(xml.EndElement{Name: suitesName})
 	enc.Close()
 	return buf.Bytes(), nil
+}
+
+// getSkipDetails checks data about the test suite, file and runs to determine why a given run was skipped
+// Test can be skipped due to:
+// 1. terraform test recieving an interrupt from users; all unstarted tests will be skipped
+// 2. A previous run in a file has failed, causing subsequent run blocks to be skipped
+func getSkipDetails(runIndex int, file *moduletest.File, suiteStopped bool) (string, string) {
+	if suiteStopped {
+		// Test suite experienced an interrupt
+		// This block only handles graceful Stop interrupts, as Cancel interrupts will prevent a JUnit file being produced at all
+		message := "Test skipped due to an interrupt"
+		body := fmt.Sprintf("Terraform received an interrupt and stopped gracefully. This caused all remaining testcases to be skipped")
+
+		return message, body
+	}
+
+	if file.Status == moduletest.Error {
+		// Overall test file marked as errored in the context of a skipped test means tests have been skipped after
+		// a previous test/run blocks has errored out
+		for i := runIndex; i >= 0; i-- {
+			if file.Runs[i].Status == moduletest.Error {
+				// Skipped due to error in previous run within the file
+				message := "Test run skipped due to a previous testcase error"
+				body := fmt.Sprintf("Previous testcase %q ended in error, which caused the test remaining tests in the file to be skipped", file.Runs[i].Name)
+				return message, body
+			}
+		}
+	}
+
+	// Unhandled case: This results in <skipped></skipped> with no attributes or body
+	return "", ""
 }
 
 func suiteFilesAsSortedList(files map[string]*moduletest.File) []*moduletest.File {
