+++ internal/command/junit/junit_test.go
@@ -1,6 +1,6 @@
 // Copyright (c) HashiCorp, Inc.
 // SPDX-License-Identifier: BUSL-1.1
-package junit
+package junit_test
 
 import (
 	"bytes"
@@ -8,6 +8,8 @@ import (
 	"os"
 	"testing"
 
+	"github.com/hashicorp/terraform/internal/backend/local"
+	"github.com/hashicorp/terraform/internal/command/junit"
 	"github.com/hashicorp/terraform/internal/configs/configload"
 	"github.com/hashicorp/terraform/internal/moduletest"
 )
@@ -20,12 +22,16 @@ func Test_TestJUnitXMLFile_Save(t *testing.T) {
 
 	cases := map[string]struct {
 		filename      string
+		runner        *local.TestSuiteRunner
 		suite         moduletest.Suite
 		expectedOuput []byte
 		expectError   bool
 	}{
-		"renders output indicating when tests are skipped": {
+		"<skipped> element can explain when skip is due to the runner being stopped by an interrupt": {
 			filename: "output.xml",
+			runner: &local.TestSuiteRunner{
+				Stopped: true,
+			},
 			suite: moduletest.Suite{
 				Status: moduletest.Skip,
 				Files: map[string]*moduletest.File{
@@ -36,6 +42,67 @@ func Test_TestJUnitXMLFile_Save(t *testing.T) {
 							{
 								Name:   "my_test",
 								Status: moduletest.Skip,
+							},
+						},
+					},
+				},
+			},
+			expectedOuput: []byte(`<?xml version="1.0" encoding="UTF-8"?><testsuites>
+  <testsuite name="file1.tftest.hcl" tests="1" skipped="1" failures="0" errors="0">
+    <testcase name="my_test" classname="file1.tftest.hcl">
+      <skipped message="Test skipped due to an interrupt"><![CDATA[Terraform received an interrupt and stopped gracefully. This caused all remaining testcases to be skipped]]></skipped>
+    </testcase>
+  </testsuite>
+</testsuites>`),
+		},
+		"<skipped> element can explain when skip is due to the previously errored runs/testcases in the file": {
+			filename: "output.xml",
+			runner:   &local.TestSuiteRunner{},
+			suite: moduletest.Suite{
+				Status: moduletest.Error,
+				Files: map[string]*moduletest.File{
+					"file1.tftest.hcl": {
+						Name:   "file1.tftest.hcl",
+						Status: moduletest.Error,
+						Runs: []*moduletest.Run{
+							{
+								Name:   "my_test_1",
+								Status: moduletest.Error,
+							},
+							{
+								Name:   "my_test_2",
+								Status: moduletest.Skip,
+							},
+						},
+					},
+				},
+			},
+			expectedOuput: []byte(`<?xml version="1.0" encoding="UTF-8"?><testsuites>
+  <testsuite name="file1.tftest.hcl" tests="2" skipped="1" failures="0" errors="1">
+    <testcase name="my_test_1" classname="file1.tftest.hcl">
+      <error message="Encountered an error"></error>
+    </testcase>
+    <testcase name="my_test_2" classname="file1.tftest.hcl">
+      <skipped message="Test run skipped due to a previous testcase error"><![CDATA[Previous testcase "my_test_1" ended in error, which caused the test remaining tests in the file to be skipped]]></skipped>
+    </testcase>
+  </testsuite>
+</testsuites>`),
+		},
+		"<skipped> element is present without additional details when contextual data is not available": {
+			filename: "output.xml",
+			runner:   &local.TestSuiteRunner{
+				// No data about being stopped
+			},
+			suite: moduletest.Suite{
+				Status: moduletest.Pending,
+				Files: map[string]*moduletest.File{
+					"file1.tftest.hcl": {
+						Name:   "file1.tftest.hcl",
+						Status: moduletest.Pending,
+						Runs: []*moduletest.Run{
+							{
+								Name:   "my_test",
+								Status: moduletest.Skip, // Only run present is skipped, no previous errors
 							},
 						},
 					},
@@ -60,10 +127,8 @@ func Test_TestJUnitXMLFile_Save(t *testing.T) {
 			loader, cleanup := configload.NewLoaderForTests(t)
 			defer cleanup()
 
-			j := TestJUnitXMLFile{
-				filename:     path,
-				configLoader: loader,
-			}
+			j := junit.NewTestJUnitXMLFile(path, loader)
+			j.SetTestSuiteRunner(tc.runner)
 
 			// Process data & save file
 			j.Save(&tc.suite)
@@ -80,145 +145,4 @@ func Test_TestJUnitXMLFile_Save(t *testing.T) {
 		})
 	}
 
-}
-
-func Test_TestJUnitXMLFile_save(t *testing.T) {
-
-	cases := map[string]struct {
-		filename    string
-		expectError bool
-	}{
-		"can save output to the specified filename": {
-			filename: func() string {
-				td := t.TempDir()
-				return fmt.Sprintf("%s/output.xml", td)
-			}(),
-		},
-		"returns an error when given a filename that isn't absolute or relative": {
-			filename:    "~/output.xml",
-			expectError: true,
-		},
-	}
-
-	for tn, tc := range cases {
-		t.Run(tn, func(t *testing.T) {
-			j := TestJUnitXMLFile{
-				filename: tc.filename,
-			}
-
-			xml := []byte(`<?xml version="1.0" encoding="UTF-8"?><testsuites>
-  <testsuite name="example_1.tftest.hcl" tests="1" skipped="0" failures="0" errors="0">
-    <testcase name="true_is_true" classname="example_1.tftest.hcl" time="0.005381209"></testcase>
-  </testsuite>
-</testsuites>`)
-
-			diags := j.save(xml)
-
-			if diags.HasErrors() {
-				if !tc.expectError {
-					t.Fatalf("got unexpected error: %s", diags.Err())
-				}
-				// return early if testing error case
-				return
-			}
-
-			if !diags.HasErrors() && tc.expectError {
-				t.Fatalf("expected an error but got none")
-			}
-
-			fileContent, err := os.ReadFile(tc.filename)
-			if err != nil {
-				t.Fatalf("unexpected error opening file")
-			}
-
-			if !bytes.Equal(fileContent, xml) {
-				t.Fatalf("wanted XML:\n%s\n got XML:\n%s\n", string(xml), string(fileContent))
-			}
-		})
-	}
-}
-
-func Test_suiteFilesAsSortedList(t *testing.T) {
-	cases := map[string]struct {
-		Suite         *moduletest.Suite
-		ExpectedNames map[int]string
-	}{
-		"no test files": {
-			Suite: &moduletest.Suite{},
-		},
-		"3 test files ordered in map": {
-			Suite: &moduletest.Suite{
-				Status: moduletest.Skip,
-				Files: map[string]*moduletest.File{
-					"test_file_1.tftest.hcl": {
-						Name:   "test_file_1.tftest.hcl",
-						Status: moduletest.Skip,
-						Runs:   []*moduletest.Run{},
-					},
-					"test_file_2.tftest.hcl": {
-						Name:   "test_file_2.tftest.hcl",
-						Status: moduletest.Skip,
-						Runs:   []*moduletest.Run{},
-					},
-					"test_file_3.tftest.hcl": {
-						Name:   "test_file_3.tftest.hcl",
-						Status: moduletest.Skip,
-						Runs:   []*moduletest.Run{},
-					},
-				},
-			},
-			ExpectedNames: map[int]string{
-				0: "test_file_1.tftest.hcl",
-				1: "test_file_2.tftest.hcl",
-				2: "test_file_3.tftest.hcl",
-			},
-		},
-		"3 test files unordered in map": {
-			Suite: &moduletest.Suite{
-				Status: moduletest.Skip,
-				Files: map[string]*moduletest.File{
-					"test_file_3.tftest.hcl": {
-						Name:   "test_file_3.tftest.hcl",
-						Status: moduletest.Skip,
-						Runs:   []*moduletest.Run{},
-					},
-					"test_file_1.tftest.hcl": {
-						Name:   "test_file_1.tftest.hcl",
-						Status: moduletest.Skip,
-						Runs:   []*moduletest.Run{},
-					},
-					"test_file_2.tftest.hcl": {
-						Name:   "test_file_2.tftest.hcl",
-						Status: moduletest.Skip,
-						Runs:   []*moduletest.Run{},
-					},
-				},
-			},
-			ExpectedNames: map[int]string{
-				0: "test_file_1.tftest.hcl",
-				1: "test_file_2.tftest.hcl",
-				2: "test_file_3.tftest.hcl",
-			},
-		},
-	}
-
-	for tn, tc := range cases {
-		t.Run(tn, func(t *testing.T) {
-			list := suiteFilesAsSortedList(tc.Suite.Files)
-
-			if len(tc.ExpectedNames) != len(tc.Suite.Files) {
-				t.Fatalf("expected there to be %d items, got %d", len(tc.ExpectedNames), len(tc.Suite.Files))
-			}
-
-			if len(tc.ExpectedNames) == 0 {
-				return
-			}
-
-			for k, v := range tc.ExpectedNames {
-				if list[k].Name != v {
-					t.Fatalf("expected element %d in sorted list to be named %s, got %s", k, v, list[k].Name)
-				}
-			}
-		})
-	}
 }
